Module rust_tcp_io_perf::connection

Functions

client_connect

close_connection

receive_message
Reads n_bytes into rbuf from the given stream. Make sure rbuf.len >= n_bytes

send_message
Sends first n_bytes from wbuf using the given stream. Make sure wbuf.len >= n_bytes

server_listen_and_get_first_connection
Starts listening on given port and return first connection to that port as a stream.

setup
Setup the streams and eventually pins the thread according to the configuration.